refactor: buildQuestionDefsをlayout.tsに移動
純粋なデータ変換ロジックをmain.tsからlib/layout.tsへ移動し、
LayoutMeta→QuestionDef[]の変換を適切なモジュールに配置。
